requirements.txt
flask
psycopg2
Flask-SQLAlchemy
Flask-Migrate
Flask-Marshmallow
marshmallow-sqlalchemy
connexion[swagger-ui]
gunicorn
pytest-flask-sqlalchemy
